+++ libraries/common/src/main/java/androidx/media3/common/DeviceInfo.java
@@ -19,9 +19,12 @@ import static java.lang.annotation.ElementType.TYPE_USE;
 
 import android.os.Bundle;
 import androidx.annotation.IntDef;
+import androidx.annotation.IntRange;
 import androidx.annotation.Nullable;
+import androidx.media3.common.util.Assertions;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -45,22 +48,83 @@ public final class DeviceInfo implements Bundleable {
   public static final int PLAYBACK_TYPE_REMOTE = 1;
 
   /** Unknown DeviceInfo. */
-  public static final DeviceInfo UNKNOWN =
-      new DeviceInfo(PLAYBACK_TYPE_LOCAL, /* minVolume= */ 0, /* maxVolume= */ 0);
+  public static final DeviceInfo UNKNOWN = new Builder(PLAYBACK_TYPE_LOCAL).build();
+
+  /** Builder for {@link DeviceInfo}. */
+  public static final class Builder {
+
+    private final @PlaybackType int playbackType;
+
+    private int minVolume;
+    private int maxVolume;
+
+    /**
+     * Creates the builder.
+     *
+     * @param playbackType The {@link PlaybackType}.
+     */
+    public Builder(@PlaybackType int playbackType) {
+      this.playbackType = playbackType;
+    }
+
+    /**
+     * Sets the minimum supported device volume.
+     *
+     * <p>The minimum will be set to {@code 0} if not specified.
+     *
+     * @param minVolume The minimum device volume.
+     * @return This builder.
+     */
+    @CanIgnoreReturnValue
+    public Builder setMinVolume(@IntRange(from = 0) int minVolume) {
+      this.minVolume = minVolume;
+      return this;
+    }
+
+    /**
+     * Sets the maximum supported device volume.
+     *
+     * @param maxVolume The maximum device volume, or {@code 0} to leave the maximum unspecified.
+     * @return This builder.
+     */
+    @CanIgnoreReturnValue
+    public Builder setMaxVolume(@IntRange(from = 0) int maxVolume) {
+      this.maxVolume = maxVolume;
+      return this;
+    }
+
+    /** Builds the {@link DeviceInfo}. */
+    public DeviceInfo build() {
+      Assertions.checkArgument(minVolume <= maxVolume);
+      return new DeviceInfo(this);
+    }
+  }
 
   /** The type of playback. */
   public final @PlaybackType int playbackType;
   /** The minimum volume that the device supports. */
+  @IntRange(from = 0)
   public final int minVolume;
-  /** The maximum volume that the device supports. */
+  /** The maximum volume that the device supports, or {@code 0} if unspecified. */
+  @IntRange(from = 0)
   public final int maxVolume;
 
-  /** Creates device information. */
+  /**
+   * @deprecated Use {@link Builder} instead.
+   */
   @UnstableApi
-  public DeviceInfo(@PlaybackType int playbackType, int minVolume, int maxVolume) {
-    this.playbackType = playbackType;
-    this.minVolume = minVolume;
-    this.maxVolume = maxVolume;
+  @Deprecated
+  public DeviceInfo(
+      @PlaybackType int playbackType,
+      @IntRange(from = 0) int minVolume,
+      @IntRange(from = 0) int maxVolume) {
+    this(new Builder(playbackType).setMinVolume(minVolume).setMaxVolume(maxVolume));
+  }
+
+  private DeviceInfo(Builder builder) {
+    this.playbackType = builder.playbackType;
+    this.minVolume = builder.minVolume;
+    this.maxVolume = builder.maxVolume;
   }
 
   @Override
@@ -96,9 +160,15 @@ public final class DeviceInfo implements Bundleable {
   @Override
   public Bundle toBundle() {
     Bundle bundle = new Bundle();
-    bundle.putInt(FIELD_PLAYBACK_TYPE, playbackType);
-    bundle.putInt(FIELD_MIN_VOLUME, minVolume);
-    bundle.putInt(FIELD_MAX_VOLUME, maxVolume);
+    if (playbackType != PLAYBACK_TYPE_LOCAL) {
+      bundle.putInt(FIELD_PLAYBACK_TYPE, playbackType);
+    }
+    if (minVolume != 0) {
+      bundle.putInt(FIELD_MIN_VOLUME, minVolume);
+    }
+    if (maxVolume != 0) {
+      bundle.putInt(FIELD_MAX_VOLUME, maxVolume);
+    }
     return bundle;
   }
 
@@ -110,6 +180,9 @@ public final class DeviceInfo implements Bundleable {
             bundle.getInt(FIELD_PLAYBACK_TYPE, /* defaultValue= */ PLAYBACK_TYPE_LOCAL);
         int minVolume = bundle.getInt(FIELD_MIN_VOLUME, /* defaultValue= */ 0);
         int maxVolume = bundle.getInt(FIELD_MAX_VOLUME, /* defaultValue= */ 0);
-        return new DeviceInfo(playbackType, minVolume, maxVolume);
+        return new DeviceInfo.Builder(playbackType)
+            .setMinVolume(minVolume)
+            .setMaxVolume(maxVolume)
+            .build();
       };
 }