+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaSessionCallbackWithMediaControllerCompatTest.java
@@ -673,8 +673,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });
@@ -702,8 +701,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });
@@ -735,8 +733,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });
@@ -761,8 +758,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });
@@ -792,8 +788,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });
@@ -819,8 +814,7 @@ public class MediaSessionCallbackWithMediaControllerCompatTest {
     handler.postAndSync(
         () -> {
           remotePlayer.deviceInfo =
-              new DeviceInfo(
-                  DeviceInfo.PLAYBACK_TYPE_REMOTE, /* minVolume= */ 0, /* maxVolume= */ 100);
+              new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_REMOTE).setMaxVolume(100).build();
           remotePlayer.deviceVolume = 23;
           session.setPlayer(remotePlayer);
         });