+++ RELEASENOTES.md
@@ -36,6 +36,7 @@
     *   Fix issue where last frame may not be rendered if the last sample with
         frames is dequeued without reading the 'end of stream' sample.
         ([#11079](https://github.com/google/ExoPlayer/issues/11079)).
+    *   Add `Builder` for `DeviceInfo` and deprecate existing constructor.
 *   Session:
     *   Deprecate 4 volume-controlling methods in `Player` and add overloaded
         methods which allow users to specify volume flags: